--- geo_analysis/location_viewer.py
@@ -1,3 +1,0 @@
-def view_locations(data):
-    for loc in data:
-        print(f"ID: {loc['id']}, Name: {loc['name']}, Region: {loc['region']}, Coordinates: {loc['coordinates']}")
